% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Irescale.R
\name{buildStabilityTable}
\alias{buildStabilityTable}
\title{Finds how many iterations are necessary to achieve stability in resampling method.}
\usage{
buildStabilityTable(data, times = 100, samples = 1000, plots = TRUE)
}
\arguments{
\item{data}{data structure after loading the file using \code{loadFile} function}

\item{times}{the number of times \code{rescaleI} will be executed. The default value is 100.}

\item{samples}{size of the resampling method. The default value is 1000}

\item{plots}{to draw the significance plot}
}
\value{
A vector with the average \eqn{\log(samples)} averages I
}
\description{
\code{buildStabilityTable}  finds how many iterations are necessary to achieve stability in resampling method, plotting in a log scale.
}
